fix(ingest/pentaho): declare extra deps, harden XML parsing, fix merge fallout
The pentaho extra previously declared no dependencies, so
`pip install acryl-datahub[pentaho]` installed neither the SQL parser nor an
XML library the source imports at module load. Add sqlglot_lib, used by
step_processors/table_input.py for SQL lineage, and defusedxml.

Replace stdlib xml.etree.ElementTree parsing with defusedxml in pentaho.py,
hardening the .ktr/.kjb parse path against XXE and billion-laughs. The Element
type is still imported from stdlib under `# nosec B405` because defusedxml does
not export it, matching the existing hana and sap-datasphere parsers.

Drop the TYPE_CHECKING guard for ProcessingContext in the step processors and
import it directly; there is no import cycle. The guard remains for
PentahoSource in base.py, where the reference is genuinely circular.

Replace SupportStatus.TESTING with ALPHA. Upstream retired the pre-2026
TESTING / CERTIFIED / INCUBATING names without aliases, so the connector no
longer imported after the merge.

Widen file_size_limit_mb from int to float so sub-MB limits are expressible;
its only consumer compares it against a float from an os.path.getsize division.

Make the get_workunits path assertion platform-agnostic via os.path.join, and
reformat the test file with the repo's pinned ruff 0.15.22.

Regenerate pyproject.toml and uv.lock from setup.py.

+++ metadata-ingestion/src/datahub/ingestion/source/pentaho/config.py
@@ -26,7 +26,7 @@ class PentahoSourceConfig(ConfigModel):
         default="pentaho_admin", description="Default owner username"
     )
 
-    file_size_limit_mb: int = Field(
+    file_size_limit_mb: float = Field(
         default=50, description="Maximum file size in MB to process"
     )
